streamlit/requirements-streamlit.txt
streamlit==1.32.0
plotly==5.18.0
pandas==2.1.4
redis==5.0.1
python-dotenv==1.0.0
confluent-kafka==2.3.0
kafka-python==2.0.2
python-json-logger==2.0.7
snowflake-connector-python
